requirements.txt
fastapi
uvicorn
openai-whisper
torch
librosa
soundfile
python-multipart
